@@ -17,7 +17,9 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 
 
-version: 1.19.3+1
+
+version: 1.20.1+1
+
 
 environment:
   sdk: ">=3.0.0 <4.0.0"
@@ -27,7 +29,7 @@ dependencies:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  intl: ^0.18.0
+  intl: ^0.19.0
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
 
@@ -51,7 +53,7 @@ dependencies:
   page_transition: ^2.0.9
   fluttertoast: ^8.0.8
   #modal_progress_hud: ^0.1.3
-  flutter_inappwebview: ^5.7.2+3
+  flutter_inappwebview: ^6.1.5
 
   # CONNECTIVITY
   http: ^0.13.6
@@ -61,12 +63,11 @@ dependencies:
   cached_network_image: ^3.2.3
   xml_parser: ^1.0.0
 
-  share_plus: ^6.3.1
-  launch_review: ^3.0.1
+  share_plus: ^10.0.0
   url_launcher: ^6.1.11
   store_redirect: ^2.0.1
   connectivity_plus: ^4.0.0
-  uni_links: ^0.5.1
+  uni_links3: ^0.5.3
 
   # TOOLS
   google_fonts: ^4.0.4
@@ -101,7 +102,7 @@ dependencies:
   rive_splash_screen: ^0.1.1
   lottie: ^2.3.2
   flutter_svg: ^2.0.5
-  dots_indicator: ^3.0.0
+  dots_indicator: 3.0.0
   sizer: ^2.0.15
 
   timeago_flutter: ^3.4.0
@@ -109,12 +110,16 @@ dependencies:
   wifi_scan: ^0.4.1
 
   #  media
-  youtube_player_flutter: ^8.1.2
+  youtube_player_flutter: ^9.1.1
   youtube_explode_dart: ^1.12.4
-  unique_identifier: ^0.3.0
+  unique_identifier: 0.4.0
   collection: any
-  disk_space: ^0.2.1
+  disk_space_plus: 0.2.4
 
+  screen_control:
+    git:
+      url: https://github.com/mawaqit/native_screen_lock_mawaqit_devices
+      ref: master
   equatable: ^2.0.5
   just_audio: ^0.9.37
 
@@ -128,7 +133,7 @@ dependencies:
 
   # internet connectivity checker
   internet_connection_checker_plus: ^1.0.1
-  flutter_kurdish_localization: ^1.0.8
+  #  flutter_kurdish_localization: ^1.0.8
   scrollview_observer: ^1.21.0
   # manage files and folders
   flutter_archive: ^6.0.3
@@ -137,6 +142,12 @@ dependencies:
 
   wakelock_plus: 1.1.4
   wakelock_plus_platform_interface: 1.1.0
+  flutter_background_service: ^5.0.9
+  notification_overlay:
+    git:
+      url: https://github.com/mawaqit/android-tv-notification
+      ref: main
+
   percent_indicator: ^4.2.3
 
   # functional programming
@@ -146,7 +157,14 @@ dependencies:
   scroll_to_index: ^3.0.1
   fast_cached_network_image: ^1.2.0
 
-  montenegrin_localization: ^0.1.0+1
+#  montenegrin_localization: ^0.1.0+1
+  permission_handler: ^10.2.0
+  workmanager: ^0.5.1
+  package_info_plus: ^4.2.0
+
+  # Added screen control package
+  flutter_screen_controller: ^0.0.6
+
 
 dev_dependencies:
   flutter_test:
@@ -157,17 +175,18 @@ dev_dependencies:
   dependency_validator:
   build_runner: ^2.4.9
   http_mock_adapter: ^0.6.1
-  flutter_background_service: ^2.1.3
   flutter_lints: 3.0.1
   hive_generator: ^2.0.1
   mocktail: ^1.0.3
   dart_mappable_builder: ^4.2.3
   json_serializable: ^6.8.0
+  fast_flutter_driver: ^3.0.0+1
+
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
 dependency_overrides:
-  win32: 5.0.5
+#  win32: 5.0.9
   rive: ^0.9.1
 # The following section is specific to Flutter.
 flutter:
