@@ -4,7 +4,6 @@ import 'package:mawaqit/i18n/AppLanguage.dart';
 import 'package:mawaqit/src/helpers/AppRouter.dart';
 import 'package:mawaqit/src/services/mosque_manager.dart';
 import 'package:mawaqit/src/widgets/MawaqitDrawer.dart';
-import 'package:mawaqit/src/widgets/MawaqitWebViewWidget.dart';
 import 'package:provider/provider.dart';
 
 /// Javascript version of the app
@@ -37,9 +36,7 @@ class _HomeScreenState extends State<HomeScreen> {
       child: Scaffold(
         key: _scaffoldKey,
         drawer: MawaqitDrawer(goHome: () => AppRouter.pop()),
-        body: MawaqitWebViewWidget(
-          path: context.watch<MosqueManager>().buildUrl(appLanguage.appLocal.languageCode),
-        ),
+        body: Container(),
       ),
     );
   }
